requirements.txt
python-dotenv
pyTelegramBotAPI
requests
Pillow
